ExecutionZoneRestController split up in multiple different controllers. URLMapping adapted. AccessService check improved. HTTP methods adapted - all resources updates are now using PUT instead of POST.
