+++ grails-app/controllers/org/zenboot/portal/processing/ExecutionZoneRestController.groovy
@@ -17,12 +17,6 @@ import org.springframework.context.ApplicationEventPublisherAware
 import org.springframework.http.HttpStatus
 import org.zenboot.portal.AbstractRestController
 import org.zenboot.portal.ControllerUtils
-import org.zenboot.portal.Customer
-import org.zenboot.portal.Host
-import org.zenboot.portal.HostState
-import org.zenboot.portal.NotificationType
-import org.zenboot.portal.ServiceUrl
-import org.zenboot.portal.UserNotification
 import org.zenboot.portal.security.Person
 import org.zenboot.portal.security.Role
 import org.zenboot.portal.Template
@@ -36,9 +30,8 @@ class ExecutionZoneRestController extends AbstractRestController implements Appl
     def grailsLinkGenerator
     def applicationEventPublisher
 
-    static allowedMethods = [index: "GET" , help: "GET", list: "GET", execute: "POST", listparams: "GET", listactions: "GET", createzone: "POST", exectypes: "GET", execzonetemplate: "GET",
-    cloneexecutionzone: "GET", listhosts: "GET", listhoststates: "GET", changeExecutionZoneParams: ["POST", "DELETE"], changeExecutionZoneAttributes: ["POST"],
-    listDetailedActions: "GET", listServiceUrls: "GET", listCustomers: "GET", listUserNotifications: "GET", editUserNotification: "POST",createUserNotification: "POST"]
+    static allowedMethods = [index: "GET" , help: "GET", list: "GET", execute: "POST", listparams: "GET", listactions: "GET", createzone: "POST", execzonetemplate: "GET",
+    cloneexecutionzone: "POST", changeExecutionZoneParams: ["PUT", "DELETE"], changeExecutionZoneAttributes: ["PUT"]]
 
     @Override
     void setApplicationEventPublisher(ApplicationEventPublisher eventPublisher) {
@@ -574,8 +567,7 @@ class ExecutionZoneRestController extends AbstractRestController implements Appl
             return
         }
 
-        if (SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN) || userHasAccess(executionZone)) {
-
+        if (accessService.userHasAccess(executionZone)) {
             File stackDir = new File(scriptDirectoryService.getZenbootScriptsDir().getAbsolutePath()
                     + "/" + executionZone.type.name + "/scripts/" + actionName)
 
@@ -656,12 +648,7 @@ class ExecutionZoneRestController extends AbstractRestController implements Appl
             return
         }
 
-        if (SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN)) {
-            scriptDir = new File(scriptDirectoryService.getZenbootScriptsDir().getAbsolutePath()
-                    + "/" + executionZone.type.name + "/scripts/" )
-        }
-        else if (userHasAccess(executionZone)) {
-
+        if (accessService.userHasAccess(executionZone)) {
             scriptDir = new File(scriptDirectoryService.getZenbootScriptsDir().getAbsolutePath()
                     + "/" + executionZone.type.name + "/scripts/" )
         }
@@ -701,32 +688,7 @@ class ExecutionZoneRestController extends AbstractRestController implements Appl
         }
     }
 
-    /**
-     * execTypes returns a list of all existing executionZoneTypes.
-     */
-    def exectypes = {
-        withFormat {
-            xml {
-                def builder = new StreamingMarkupBuilder()
-                builder.encoding = 'UTF-8'
 
-                String executionZoneTypes = builder.bind {
-                    executionZoneTypes {
-                        ExecutionZoneType.list().sort().each {
-                            executionZoneType it
-                        }
-                    }
-                }
-
-                def xml = XmlUtil.serialize(executionZoneTypes).replace('<?xml version="1.0" encoding="UTF-8"?>', '<?xml version="1.0" encoding="UTF-8"?>\n')
-                xml = xml.replaceAll('<([^/]+?)/>', '<$1></$1>')
-                render contentType: "text/xml", xml
-            }
-            json {
-                render (contentType: "text/json") { ExecutionZoneType.list().sort() } as JSON
-            }
-        }
-    }
 
     /**
      * This method returns a xml or json template to create an execution zone.
@@ -1194,7 +1156,7 @@ class ExecutionZoneRestController extends AbstractRestController implements Appl
             return
         }
 
-        if (SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN) || userHasAccess(executionZone)) {
+        if (accessService.userHasAccess(executionZone)) {
 
             execution.each {
                 if (it.any { key, value -> value == '' || value == null}) {
@@ -1321,161 +1283,6 @@ class ExecutionZoneRestController extends AbstractRestController implements Appl
     }
 
     /**
-     * The method returns the hosts. The result could be more specific if 'hostState' parameter is added to the request url e.g. ?hostState=completed to return all
-     * hosts with the state completed. It is also possible to add multiple states. In this case call the url with ?hostState=completed,created . Furthermore it is possible
-     * to restrict the search for a single execution zone. In this case add e.g. ?execId=104 to the url. You also can use both e.g. ?execId=104&hostState=completed,created .
-     */
-    def listhosts = {
-            ExecutionZone executionZone
-
-            if (params.execId && params.execId.isInteger()) {
-                if (ExecutionZone.findById(params.execId as Long)) {
-                    executionZone = ExecutionZone.findById(params.execId as Long)
-                } else {
-                    this.renderRestResult(HttpStatus.NOT_FOUND, null, null, 'ExecutionZone with id ' + params.execId + ' not found.')
-                    return
-                }
-            }
-
-            List<Long> usersExecutionZones = []
-            if (!SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN)) {
-                Map execZoneMap = accessService.accessCache[springSecurityService.getCurrentUserId()]
-                usersExecutionZones = execZoneMap.findAll{it.value == true}.collect{it.key}
-                if (usersExecutionZones.isEmpty()) {
-                    this.renderRestResult(HttpStatus.UNAUTHORIZED, null, null, 'You do not have any access to any executionzone.')
-                }
-            }
-
-            def hostsFromZone = []
-
-            if (params.hostState) {
-                def hostStates = []
-
-                if (params.hostState.contains(',')) {
-                    hostStates = params.hostState.split(',')
-                } else {
-                    hostStates.add(params.hostState as String)
-                }
-
-                hostStates.each {
-                    String state = it as String
-                    state = state.toUpperCase()
-                    if (HostState.values().find { it.toString() == state }) {
-                        HostState hostState = HostState.valueOf(state)
-                        if (executionZone) {
-                            if (SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN) || userHasAccess(executionZone)) {
-                                hostsFromZone.addAll(Host.findAllByExecZoneAndState(executionZone, hostState))
-                            }
-                            else {
-                                this.renderRestResult(HttpStatus.UNAUTHORIZED, null, null, 'You have no permissions to list the hosts of this execution zone.')
-                            }
-                        } else {
-                            if (SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN)) {
-                                hostsFromZone.addAll(Host.findAllByState(hostState))
-                            }
-                            else {
-                                usersExecutionZones.each {
-                                    hostsFromZone.addAll(Host.findAllByStateAndExecZone(hostState, ExecutionZone.findById(it)))
-                                }
-                            }
-                        }
-                    } else {
-                        this.renderRestResult(HttpStatus.NOT_FOUND, null, null, 'No hoststate found for state: ' + params.hostState)
-                        return
-                    }
-                }
-            } else {
-                if (executionZone) {
-                    if (SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN) || userHasAccess(executionZone)) {
-                        hostsFromZone = Host.findAllByExecZone(executionZone)
-                    }
-                    else {
-                        this.renderRestResult(HttpStatus.UNAUTHORIZED, null, null, 'You have no permissions to list the hosts of this execution zone.')
-                    }
-                } else {
-                    if (SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN)) {
-                        hostsFromZone = Host.list()
-                    }
-                    else {
-                        usersExecutionZones.each {
-                            hostsFromZone.addAll(Host.findAllByExecZone(ExecutionZone.findById(it)))
-                        }
-                    }
-                }
-            }
-
-            withFormat {
-                xml {
-                    def builder = new StreamingMarkupBuilder()
-                    builder.encoding = 'UTF-8'
-
-                    String foundHosts = builder.bind {
-                        hosts {
-                            hostsFromZone.each { hostElement ->
-                                host {
-                                    hostname hostElement.hostname.toString()
-                                    cname hostElement.cname
-                                    hoststate hostElement.state.toString()
-                                    ipadress hostElement.ipAddress
-                                    serviceUrls {
-                                        hostElement.serviceUrls.each { singleurl ->
-                                            serviceUrl singleurl.url
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-
-                    def xml = XmlUtil.serialize(foundHosts).replace('<?xml version="1.0" encoding="UTF-8"?>', '<?xml version="1.0" encoding="UTF-8"?>\n')
-                    xml = xml.replaceAll('<([^/]+?)/>', '<$1></$1>')
-                    render contentType: "text/xml", xml
-                }
-                json {
-                    Map hosts = [:]
-                    List host = hostsFromZone.collect {
-                        [hostname: it.hostname.toString(), cname: it.cname, hoststate: it.state.toString(), ipadress: it.ipAddress, serviceUrls: [it.serviceUrls.collect {
-                            it.url
-                        }]]
-                    }
-                    hosts.put('hosts', host)
-                    render hosts as JSON
-                }
-            }
-    }
-
-    /**
-     * The method returns a list of all existing states of a host.
-     */
-    def listhoststates = {
-        def hostStates = HostState.findAll().collect{it.toString()}
-
-        withFormat {
-            xml {
-                def builder = new StreamingMarkupBuilder()
-                builder.encoding = 'UTF-8'
-
-                String states = builder.bind {
-                    hoststates {
-                        hostStates.each {
-                            hoststate it
-                        }
-                    }
-                }
-
-                def xml = XmlUtil.serialize(states).replace('<?xml version="1.0" encoding="UTF-8"?>', '<?xml version="1.0" encoding="UTF-8"?>\n')
-                xml = xml.replaceAll('<([^/]+?)/>', '<$1></$1>')
-                render contentType: "text/xml", xml
-            }
-            json {
-                Map jsonhoststates = [:]
-                jsonhoststates.put('hoststates', hostStates)
-                render jsonhoststates as JSON
-            }
-        }
-    }
-
-    /**
      * The method changes the processing parameters of an existing execution zone. It is possible to change / add processing parameters due add
      * a new key / value parameter pair in the data or change the value of an existing one. To change / add processing parameters you have to use
      * request method POST. If you want to remove processing parameters you have to use the request method DELETE. Keep in mind that all key / value
@@ -1494,7 +1301,7 @@ class ExecutionZoneRestController extends AbstractRestController implements Appl
         if (params.execId) {
             zone = ExecutionZone.get(params.execId as Long)
             if (zone) {
-                if (!userHasAccess(zone)) {
+                if (!accessService.userHasAccess(zone)) {
                     this.renderRestResult(HttpStatus.FORBIDDEN, null, null, 'You have no permissions to change the parameters ' +
                             'of the executionZone with id ' + params.execId + '.')
                     return
@@ -1566,7 +1373,7 @@ class ExecutionZoneRestController extends AbstractRestController implements Appl
         Set<ProcessingParameter> new_parameters = [] as SortedSet
         Set<ProcessingParameter> validation = [] as SortedSet
 
-        if (request.method == 'POST') {
+        if (request.method == 'PUT') {
             parameters.each { key, value ->
 
                 ProcessingParameter parameter = origin_parameters.find { it.name == key }
@@ -1643,7 +1450,7 @@ class ExecutionZoneRestController extends AbstractRestController implements Appl
             if (params.execId) {
                 zone = ExecutionZone.get(params.execId as Long)
                 if (zone) {
-                    if (!userHasAccess(zone)) {
+                    if (!accessService.userHasAccess(zone)) {
                         this.renderRestResult(HttpStatus.FORBIDDEN, null, null, 'You have no permissions to change the parameters ' +
                                 'of the executionZone with id ' + params.execId + '.')
                         return
@@ -1736,7 +1543,7 @@ class ExecutionZoneRestController extends AbstractRestController implements Appl
             }
 
             if (!zone.save(flush: true)) {
-                this.renderRestResult(HttpStatus.CONFLICT, null, null, 'An error occurred while updating the execution zone.')
+                this.renderRestResult(HttpStatus.INTERNAL_SERVER_ERROR, null, null, 'An error occurred while updating the execution zone.')
                 return
             }
 
@@ -1749,719 +1556,6 @@ class ExecutionZoneRestController extends AbstractRestController implements Appl
         else {
             this.renderRestResult(HttpStatus.UNAUTHORIZED, null, null, 'You have no permissions to change the attributes of an execution zone.')
         }
-    }
-
-    /**
-     * The method returns a detailed list of execution zone actions. It is possible to specify the execution zone or a list of execution zones delimited by ',' (?execId=1,2,3,5...).
-     * If you don't specify the execId the method tries to return all execution zone actions as once. If the number of results is bigger than 100, the pagination mechanism will be forced
-     * so the method returns a list of prepared urls.
-     */
-    def listDetailedActions = {
-        List<ExecutionZone> execZones = []
-        int offset
-        if (params.execId) {
-            if (params.execId.contains(',')) {
-                List<String> execZoneIds = params.execId.split(',')
-                execZoneIds.each {
-                    if(it.isInteger()) {
-                        ExecutionZone zone = ExecutionZone.get(it as Long)
-                        if (zone) {
-                            if (SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN) || userHasAccess(zone)) {
-                                execZones.add(zone)
-                            } else {
-                                this.renderRestResult(HttpStatus.FORBIDDEN, null, null, 'You have no permissions to get the actions ' +
-                                        'of the executionZone with id ' + params.execId + '.')
-                                return
-                            }
-                        } else {
-                            this.renderRestResult(HttpStatus.NOT_FOUND, null, null, 'No execution zone with id ' + it + ' found.')
-                            return
-                        }
-                    }
-                    else {
-                        this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, 'The execId param ' + it +' invalid. ' +
-                                'It has to be a Long value.')
-                        return
-                    }
-                }
-            } else if (params.execId.isInteger()) {
-                ExecutionZone zone = ExecutionZone.get(params.execId as Long)
-                if (zone) {
-                    if (SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN) || userHasAccess(zone)){
-                        execZones.add(zone)
-                    }
-                    else {
-                        this.renderRestResult(HttpStatus.FORBIDDEN, null, null, 'You have no permissions to change the parameters ' +
-                                'of the executionZone with id ' + params.execId + '.')
-                        return
-                    }
-                } else {
-                    this.renderRestResult(HttpStatus.NOT_FOUND, null, null, 'No execution zone with id ' + params.execId + ' found.')
-                    return
-                }
-            } else {
-                this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, 'The execId param invalid. ' +
-                        'It has to be a Long value or a list of Long values delimited by ","')
-                return
-            }
-        } else {
-            if (SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN)) {
-                execZones = ExecutionZone.getAll()
-            } else {
-                def execZoneIds = accessService.accessCache[springSecurityService.getCurrentUserId() as Long]?
-                        accessService.accessCache[springSecurityService.getCurrentUserId() as Long].findAll {it.value == true}.collect {it.key}
-                        : []
-
-                execZoneIds.each {
-                    execZones.add(ExecutionZone.get(it as Long))
-                }
-            }
-        }
-
-        if (params.offset && params.offset.isInteger()) {
-            offset = params.offset as Integer
-        }
-        else {
-            offset = 0
-        }
-
-        withFormat {
-            xml {
-                def builder = new StreamingMarkupBuilder()
-                builder.encoding = 'UTF-8'
-
-                int totalSize = execZones.sum {it.actions.size()} as Integer
-
-                if (100 < totalSize && offset == 0) {
-                    String actionsXML = builder.bind {
-                        pagination {
-                            execZones.each { ExecutionZone execZone ->
-                                int start = 0
-                                def zoneSize = execZone.actions.size()
-                                def pages = Math.round((zoneSize / 100) + 0.5) //round to next full page
-
-                                pages.times {
-                                    url "/zenboot/rest/v1/listactions?execId=" + execZone.id + "&offset=" + start
-
-                                    if ((start + 100) < zoneSize) {
-                                        start += 100
-                                    }
-                                }
-                            }
-                        }
-                    }
-
-                    def xml = XmlUtil.serialize(actionsXML).replace('<?xml version="1.0" encoding="UTF-8"?>', '<?xml version="1.0" encoding="UTF-8"?>\n')
-                    xml = xml.replaceAll('<([^/]+?)/>', '<$1></$1>')
-                    render contentType: "text/xml", xml
-                }
-                else {
-                    String actionsXML = builder.bind {
-                        executionZoneActions {
-                            execZones.each { execZone ->
-                               def list = execZone.actions as List
-                                if (list.size() > (offset+100)) {
-                                    list = list.subList(offset, offset + 100)
-                                } else {
-                                    list = list.subList(offset, (list.size()-1))
-                                }
-                                list.each { ExecutionZoneAction action ->
-                                    executionZoneAction {
-                                        execId execZone.id
-                                        execActionId action.id
-                                        execActionName action.toString()
-                                        scriptDir action.scriptDir.path
-                                        creationDate action.creationDate.toString()
-                                        processingParameters {
-                                            action.processingParameters.each { ProcessingParameter pp ->
-                                                processingParameterName pp.name
-                                                processingParameterValue pp.value
-                                            }
-                                        }
-                                        scriptletBatchIds action.scriptletBatches.collect { it.id }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    def xml = XmlUtil.serialize(actionsXML).replace('<?xml version="1.0" encoding="UTF-8"?>', '<?xml version="1.0" encoding="UTF-8"?>\n')
-                    xml = xml.replaceAll('<([^/]+?)/>', '<$1></$1>')
-                    render contentType: "text/xml", xml
-                }
-            }
-            json {
-                int totalSize = execZones.sum { it.actions.size() } as Integer
-
-                if (100 < totalSize && offset == 0) {
-                    def paginationJSON = []
-                    execZones.each { ExecutionZone execZone ->
-                        int start = 0
-                        def zoneSize = execZone.actions.size()
-                        def pages = Math.round((zoneSize / 100) + 0.5) //round to next full page
-
-                        pages.times {
-                            paginationJSON.add("/zenboot/rest/v1/listactions?execId=" + execZone.id + "&offset=" + start)
-
-                            if ((start + 100) < zoneSize) {
-                                start += 100
-                            }
-                        }
-                    }
-                    render paginationJSON as JSON
-                } else {
-
-                    def actionsJSON = [:]
-                    def actionsList = []
-                    execZones.each { ExecutionZone execZone ->
-                        def list = execZone.actions as List
-                        if (list.size() > (offset+100)) {
-                            list = list.subList(offset, offset + 100)
-                        } else {
-                            list = list.subList(offset, (list.size()-1))
-                        }
-                        list.each { ExecutionZoneAction action ->
-                            def actionJSON = [:]
-                            actionJSON.put('execId', execZone.id)
-                            actionJSON.put('execActionId', action.id)
-                            actionJSON.put('execActionName', action.toString())
-                            actionJSON.put('scriptDir', action.scriptDir.path)
-                            actionJSON.put('creationDate', action.creationDate.toString())
-                            def processingParameters = [:]
-                            action.processingParameters.each { ProcessingParameter pp ->
-                                processingParameters.put(pp.name, pp.value)
-                            }
-                            actionJSON.put('processingParameters', processingParameters)
-                            actionJSON.put('scriptletBatchIds', action.scriptletBatches.collect { it.id })
-                            actionsList.add(actionJSON)
-                        }
-                    }
-                    actionsJSON.put('executionZoneActions', actionsList)
-                    render actionsJSON as JSON
-                }
-            }
-        }
-    }
-
-    /**
-     * The method return a list with active hosts service urls. It is possible to specify the execution zone or a list of execution zones delimited by ',' (?execId=1,2,3,5...). If the execId param is
-     * not set, the method returns the service urls of all execution zones.
-     */
-    def listServiceUrls = {
-        def execZones = []
-        if (params.execId) {
-            if (params.execId.contains(',')) {
-                List<String> execZoneIds = params.execId.split(',')
-                execZoneIds.each {
-                    if(it.isInteger()) {
-                        ExecutionZone zone = ExecutionZone.get(it as Long)
-                        if (zone) {
-                            if (SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN) || userHasAccess(zone)) {
-                                execZones.add(zone)
-                            } else {
-                                this.renderRestResult(HttpStatus.FORBIDDEN, null, null, 'You have no permissions to change the parameters ' +
-                                        'of the executionZone with id ' + params.execId + '.')
-                                return
-                            }
-                        } else {
-                            this.renderRestResult(HttpStatus.NOT_FOUND, null, null, 'No execution zone with id ' + it + ' found.')
-                            return
-                        }
-                    }
-                    else {
-                        this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, 'The execId param ' + it +' invalid. ' +
-                                'It has to be a Long value.')
-                        return
-                    }
-                }
-            } else if (params.execId.isInteger()) {
-                ExecutionZone zone = ExecutionZone.get(params.execId as Long)
-                if (zone) {
-                    if (SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN) || userHasAccess(zone)){
-                        execZones.add(zone)
-                    }
-                    else {
-                        this.renderRestResult(HttpStatus.FORBIDDEN, null, null, 'You have no permissions to get the serive urls ' +
-                                'of the executionZone with id ' + params.execId + '.')
-                        return
-                    }
-                } else {
-                    this.renderRestResult(HttpStatus.NOT_FOUND, null, null, 'No execution zone with id ' + params.execId + ' found.')
-                    return
-                }
-            } else {
-                this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, 'The execId param invalid. ' +
-                        'It has to be a Long value or a list of Long values delimited by ","')
-                return
-            }
-        } else {
-            if (SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN)) {
-                execZones = ExecutionZone.getAll()
-            } else {
-                def execZoneIds = accessService.accessCache[springSecurityService.getCurrentUserId() as Long]?
-                        accessService.accessCache[springSecurityService.getCurrentUserId() as Long].findAll {it.value == true}.collect {it.key}
-                        : []
-
-                execZoneIds.each {
-                    execZones.add(ExecutionZone.get(it as Long))
-                }
-            }
-        }
-
-        withFormat {
-            xml {
-                def builder = new StreamingMarkupBuilder()
-                builder.encoding = 'UTF-8'
-
-                String serviceURLsXML = builder.bind {
-                    serviceURLs {
-                        execZones.each { ExecutionZone zone ->
-                            zone.getActiveServiceUrls().each { ServiceUrl activeServiceUrl ->
-                                serviceURL {
-                                    id activeServiceUrl.id
-                                    url activeServiceUrl.url
-                                    serviceowner {
-                                        ipaddress activeServiceUrl.owner.ipAddress
-                                        cname activeServiceUrl.owner.cname
-                                        creationDate activeServiceUrl.owner.creationDate
-                                        expiryDate activeServiceUrl.owner.expiryDate
-                                        state activeServiceUrl.owner.state.name()
-                                        host {
-                                            id activeServiceUrl.owner.hostname.id
-                                            name activeServiceUrl.owner.hostname.name
-                                        }
-                                    }
-                                    execId zone.id
-                                }
-                            }
-                        }
-                    }
-                }
-
-                def xml = XmlUtil.serialize(serviceURLsXML).replace('<?xml version="1.0" encoding="UTF-8"?>', '<?xml version="1.0" encoding="UTF-8"?>\n')
-                xml = xml.replaceAll('<([^/]+?)/>', '<$1></$1>')
-                render contentType: "text/xml", xml
-            }
-            json {
-                def serviceURLsJSON = []
-                execZones.each { ExecutionZone zone ->
-                    zone.getActiveServiceUrls().each { ServiceUrl activeServiceUrl ->
-                        def serviceURLJSON = [:]
-                        serviceURLJSON.put('id', activeServiceUrl.id)
-                        serviceURLJSON.put('url', activeServiceUrl.url)
-                        def serviceOwner = [:]
-                        serviceOwner.put('ipaddress', activeServiceUrl.owner.ipAddress)
-                        serviceOwner.put('cname', activeServiceUrl.owner.cname)
-                        serviceOwner.put('creationDate', activeServiceUrl.owner.creationDate)
-                        serviceOwner.put('expiryDate', activeServiceUrl.owner.expiryDate)
-                        serviceOwner.put('state', activeServiceUrl.owner.state.name())
-                        def host = [:]
-                        host.put('id', activeServiceUrl.owner.hostname.id)
-                        host.put('name', activeServiceUrl.owner.hostname.name)
-                        serviceOwner.put('host', host)
-                        serviceURLJSON.put('serviceOwner', serviceOwner)
-                        serviceURLJSON.put('execId', zone.id)
-                        serviceURLsJSON.add(serviceURLJSON)
-                    }
-                }
-                render serviceURLsJSON as JSON
-            }
-        }
-    }
-
-    /**
-     * The method return a list with customers. It is possible to specify the customer by email or a list of emails delimited by ',' (?email=my.email.com,my.email2.com,...). It is also possible
-     * to get specify the customer by id or a list of ids delimted by ',' (?customerId=1,2,...).
-     * If neither email nor customerId is set, the method returns a list of all customers.
-     *
-     * Admin permissions are required.
-     */
-    def listCustomers = {
-        if (SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN)) {
-            List<Customer> customersCollection = []
-            if (params.email) {
-                if (params.email.contains(',')) {
-                    List<String> emails = params.email.split(',')
-                    emails.each {
-                        Customer customer = Customer.findByEmail(it)
-                        if (customer) {
-                            customersCollection.add(customer)
-                        } else {
-                            this.renderRestResult(HttpStatus.NOT_FOUND, null, null, 'No customer with email ' + it + ' found.')
-                            return
-                        }
-                    }
-                } else {
-                    Customer customer = Customer.findByEmail(params.email as String)
-                    if (customer) {
-                        customersCollection.add(customer)
-                    } else {
-                        this.renderRestResult(HttpStatus.NOT_FOUND, null, null, 'No customer with email ' + params.email + ' found.')
-                        return
-                    }
-                }
-            } else if (params.customerId) {
-                if (params.customerId.contains(',')) {
-                    List<String> iDs = params.customerId.split(',')
-                    iDs.each {
-                        if (it.isInteger()) {
-                            Customer customer = Customer.findById(it as Long)
-                            if (customer) {
-                                customersCollection.add(customer)
-                            } else {
-                                this.renderRestResult(HttpStatus.NOT_FOUND, null, null, 'No customer with id ' + it + ' found.')
-                                return
-                            }
-
-                        } else {
-                            this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, 'The customerId param is invalid. ' +
-                                    'It has to be a Long value or a list of Long values delimited by ","')
-                            return
-                        }
-                    }
-                } else {
-                    if (params.customerId.isInteger()) {
-                        Customer customer = Customer.findById(params.customerId as Long)
-                        if (customer) {
-                            customersCollection.add(customer)
-                        } else {
-                            this.renderRestResult(HttpStatus.NOT_FOUND, null, null, 'No customer with id ' + params.customerId + ' found.')
-                            return
-                        }
-                    } else {
-                        this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, 'The customerId param is invalid. ' +
-                                'It has to be a Long value or a list of Long values delimited by ","')
-                        return
-                    }
-                }
-            } else {
-                customersCollection.addAll(Customer.getAll())
-            }
-            
-            withFormat {
-                xml {
-                    def builder = new StreamingMarkupBuilder()
-                    builder.encoding = 'UTF-8'
-                    String customersXML = builder.bind {
-                        customers {
-                            customersCollection.each { Customer customerData ->
-                                customer {
-                                    id customerData.id
-                                    email customerData.email
-                                    creationDate customerData.creationDate
-                                    hosts {
-                                        customerData.hosts.each {Host customerHost ->
-                                            host customerHost.cname
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                    def xml = XmlUtil.serialize(customersXML).replace('<?xml version="1.0" encoding="UTF-8"?>', '<?xml version="1.0" encoding="UTF-8"?>\n')
-                    xml = xml.replaceAll('<([^/]+?)/>', '<$1></$1>')
-                    render contentType: "text/xml", xml
-                }
-                json {
-                    def customers = []
-                    customersCollection.each { Customer customerData ->
-                        def customer = [:]
-                        customer.put('id',customerData.id)
-                        customer.put('email', customerData.email)
-                        customer.put('creationDate', customerData.creationDate)
-                        def hosts = []
-                        customerData.hosts.each {Host customerHost ->
-                            hosts.add(customerHost.cname)
-                        }
-                        customer.put('hosts', hosts)
-                        customers.add(customer)
-                    }
-                    render customers as JSON
-                }
-            }
-        } else {
-            this.renderRestResult(HttpStatus.FORBIDDEN, null, null, 'Only admins are allowed to request these resources.')
-        }
-    }
-
-    /**
-     * The method return a list of user notifications. It is possible to specify the enabled param to get all enabled or disabled user notifications. If the enabled parameter is not set, the method
-     * return all available user notifications. Admin permissions are required the get the information of this resource.
-     */
-    def listUserNotifications = {
-        if (SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN)) {
-            List<UserNotification> userNotificationsData = []
-            if (params.enabled != null) {
-                userNotificationsData = UserNotification.findAllByEnabled(params.enabled.toBoolean())
-            }
-            else {
-                userNotificationsData = UserNotification.getAll()
-            }
-
-            withFormat {
-                xml {
-                    def builder = new StreamingMarkupBuilder()
-                    builder.encoding = 'UTF-8'
-                    String userNotificationsXML = builder.bind {
-                        usernotifications {
-                            userNotificationsData.each { UserNotification userNotificationData ->
-                                usernotification {
-                                    id userNotificationData.id
-                                    replacemewithmessage userNotificationData.message
-                                    notificationtype userNotificationData.type.name()
-                                    enabled userNotificationData.enabled
-                                }
-                            }
-                        }
-                    }
-                    def xml = XmlUtil.serialize(userNotificationsXML).replace('<?xml version="1.0" encoding="UTF-8"?>', '<?xml version="1.0" encoding="UTF-8"?>\n')
-                    xml = xml.replaceAll('<([^/]+?)/>', '<$1></$1>')
-                    xml = xml.replaceAll('replacemewithmessage', 'message')
-                    render contentType: "text/xml", xml
-                }
-                json {
-                    def usernotifications = []
-                    userNotificationsData.each { UserNotification userNotificationData ->
-                        def usernotification = [:]
-                        usernotification.put('id', userNotificationData.id)
-                        usernotification.put('message', userNotificationData.message)
-                        usernotification.put('notificationtype', userNotificationData.type.name())
-                        usernotification.put('enabled', userNotificationData.enabled)
-                        usernotifications.add(usernotification)
-                    }
-                    render usernotifications as JSON
-                }
-            }
-        } else {
-        this.renderRestResult(HttpStatus.FORBIDDEN, null, null, 'Only admins are allowed to request these resources.')
-        }
-
-    }
-
-    /**
-     * The method override the values of an existing user notification. Admin permissions are required to edit an user notification.
-     */
-    def editUserNotification = {
-        if (SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN)) {
-            Boolean hasError = Boolean.FALSE
-            UserNotification userNotificationData
-            if (params.notificationId.isInteger()) {
-                userNotificationData = UserNotification.findById(params.notificationId as Long)
-                if (userNotificationData) {
-                    // do nothing
-                } else {
-                    this.renderRestResult(HttpStatus.NOT_FOUND, null, null, 'No UserNotification with id ' + params.notificationId + ' found.')
-                    return
-                }
-            } else {
-                this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, 'UserNotificationId (notificationId) not set.')
-                return
-            }
-
-            request.withFormat {
-                xml {
-                    def xml
-                    try {
-                        xml = request.XML
-                    }
-                    catch (ConverterException e) {
-                        this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, e.message)
-                        hasError = Boolean.TRUE
-                        return
-                    }
-
-                    def xmlParameters = xml[0].children
-                    def parameters = [:]
-
-                    xmlParameters.each { node ->
-                        def name = ''
-                        def value = ''
-                        node.children.each { innerNode ->
-                            if (innerNode.name == 'parameterName') {
-                                name = innerNode.text()
-                            } else if (innerNode.name == 'parameterValue') {
-                                value = innerNode.text()
-                            }
-                        }
-                        parameters.put(name, value)
-                    }
-
-                    parameters.each {
-                        if (userNotificationData.hasProperty(it.key)) {
-                            userNotificationData.properties[it.key] = it.value
-                        } else {
-                            this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, 'Property ' + it.key + ' not exists for UserNotifications.')
-                            hasError = Boolean.TRUE
-                            return
-                        }
-                    }
-                }
-                json {
-                    String text = request.getReader().text
-                    def json
-
-                    try {
-                        json = new JSONObject(text)
-                    }
-                    catch (JSONException e) {
-                        this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, e.getMessage())
-                        hasError = Boolean.TRUE
-                        return
-                    }
-
-                    if (json.parameters) {
-                        json.parameters.each {
-                            if (userNotificationData.hasProperty(it.parameterName)) {
-                                userNotificationData.properties[it.parameterName] = it.parameterValue
-                            } else {
-                                this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, 'Property ' + it.parameterName + ' not exists for UserNotifications.')
-                                hasError = Boolean.TRUE
-                                return
-                            }
-                        }
-                    }
-                }
-            }
-
-            if (hasError) {
-                return
-            }
-
-            userNotificationData.save(flush: true)
-            this.renderRestResult(HttpStatus.OK, null, null, 'Notification changed.')
-        } else {
-            this.renderRestResult(HttpStatus.FORBIDDEN, null, null, 'Only admins are allowed to request these resources.')
-        }
-    }
-
-    /**
-     * The method creates a new user notification. Admin permissions are required to create a new one.
-     */
-    def createUserNotification = {
-        if (SpringSecurityUtils.ifAllGranted(Role.ROLE_ADMIN)) {
-            Boolean hasError = Boolean.FALSE
-
-            request.withFormat {
-                xml {
-                    def xml
-                    try {
-                        xml = request.XML
-                    }
-                    catch (ConverterException e) {
-                        this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, e.message)
-                        hasError = Boolean.TRUE
-                        return
-                    }
-
-                    def xmlParameters = xml[0].children
-                    def parameters = [:]
-
-                    xmlParameters.each { node ->
-                        def name = ''
-                        def value = ''
-                        node.children.each { innerNode ->
-                            if (innerNode.name == 'parameterName') {
-                                name = innerNode.text()
-                            } else if (innerNode.name == 'parameterValue') {
-                                value = innerNode.text()
-                            }
-                        }
-                        parameters.put(name, value)
-                    }
-
-                    if (parameters || parameters.any {it.value != '' || it.value != null }) {
-                        if(parameters.containsKey('enabled') && parameters.containsKey('message') && parameters.containsKey('notificationtype')) {
-                            if (NotificationType.values().any {it.name().toLowerCase() == parameters['notificationtype'].toString().toLowerCase()}) {
-                                UserNotification newUserNotification = new UserNotification(enabled: parameters['enabled'], message: parameters['message'],
-                                        type: NotificationType.valueOf(parameters['notificationtype'] as String))
-
-                                if (!newUserNotification.save(flush: true)) {
-                                    this.renderRestResult(HttpStatus.INTERNAL_SERVER_ERROR, null, null, 'An error occured while saving the new user notification. Please check your data.')
-                                    hasError = Boolean.TRUE
-                                }
-                            } else {
-                                this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, 'No type of UserNotification found for value ' + parameters['notificationtype'] +
-                                        '. Please check documentation.')
-                                hasError = Boolean.TRUE
-                                return
-                            }
-                        } else {
-                            this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, 'Parameter missing. Please check documentation.')
-                            hasError = Boolean.TRUE
-                            return
-                        }
-                    } else {
-                        this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, 'No data received or wrong data structure. Please check documentation.')
-                        hasError = Boolean.TRUE
-                        return
-                    }
-                }
-                json {
-                    String text = request.getReader().text
-                    def json
-
-                    try {
-                        json = new JSONObject(text)
-                    }
-                    catch (JSONException e) {
-                        this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, e.getMessage())
-                        hasError = Boolean.TRUE
-                        return
-                    }
-
-                    def parameters = [:]
-
-                    if (json.parameters) {
-                        json.parameters.each {
-                            parameters[it.parameterName] = it.parameterValue
-                        }
-                    }
-
-                    if (parameters || parameters.any {it.value != '' || it.value != null }) {
-                        if(parameters.containsKey('enabled') && parameters.containsKey('message') && parameters.containsKey('notificationtype')) {
-                            if (NotificationType.values().any {it.name().toLowerCase() == parameters['notificationtype'].toString().toLowerCase()}) {
-                                UserNotification newUserNotification = new UserNotification(enabled: parameters['enabled'], message: parameters['message'],
-                                        type: NotificationType.valueOf(parameters['notificationtype'] as String))
-
-                                if (!newUserNotification.save(flush: true)) {
-                                    this.renderRestResult(HttpStatus.INTERNAL_SERVER_ERROR, null, null, 'An error occured while saving the new user notification. Please check your data.')
-                                    hasError = Boolean.TRUE
-                                }
-                            } else {
-                                this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, 'No type of UserNotification found for value ' + parameters['notificationtype'] +
-                                        '. Please check documentation.')
-                                hasError = Boolean.TRUE
-                                return
-                            }
-                        } else {
-                            this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, 'Parameter missing. Please check documentation.')
-                            hasError = Boolean.TRUE
-                            return
-                        }
-                    } else {
-                        this.renderRestResult(HttpStatus.BAD_REQUEST, null, null, 'No data received or wrong data structure. Please check documentation.')
-                        hasError = Boolean.TRUE
-                    }
-                }
-            }
-            if (hasError) {
-                return
-            }
-            this.renderRestResult(HttpStatus.CREATED, null, null, 'Notification created.')
-        } else {
-            this.renderRestResult(HttpStatus.FORBIDDEN, null, null, 'Only admins are allowed to request these resources.')
-        }
-    }
-
-    /**
-     * Check if the user is already in the cache and has access to the requested execution zone.
-     * @param executionZone - the execution zone which has to be checked for access.
-     * @return true if the user has access otherwise false.
-     */
-    private Boolean userHasAccess(ExecutionZone executionZone) {
-        Long currentUserId = springSecurityService.getCurrentUserId() as Long
-        return accessService.accessCache[currentUserId] != null ?
-                accessService.accessCache[currentUserId][executionZone.id] :
-                accessService.userHasAccess(executionZone)
     }
 
     /**
